reformatting

+++ tests/test_TearDown.py
@@ -1,8 +1,14 @@
+import os
 import shutil
 import pytest
 
 
-@pytest.fixture(scope="session")
-def Testteardown():
+@pytest.fixture(scope="session", autouse=True)
+def teardown():
     yield
-    shutil.rmtree('tmp')
+    if os.path.exists('tmp'):
+        shutil.rmtree('tmp')
+    if os.path.exists('AutoClass'):
+        shutil.rmtree('AutoClass')
+    if os.path.exists('AutoReg'):
+        shutil.rmtree('AutoReg')
